Firebases Authentication and Google Sign-in are now fully implemented
