+++ lib/components/footer/bottomNavigationTabs.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:off_top_mobile/login.dart';
+import 'package:off_top_mobile/login/login.dart';
 import 'package:off_top_mobile/recordingSession.dart';
 import 'package:off_top_mobile/settings_page.dart';
