--- lib/login.dart
@@ -1,151 +1,0 @@
-import 'dart:async';
-import 'dart:convert';
-
-import 'package:flutter/material.dart';
-import 'package:off_top_mobile/recordingSession.dart';
-import 'package:off_top_mobile/routing/routing_constants.dart';
-import 'package:off_top_mobile/components/offTopTitle.dart';
-
-import 'package:http/http.dart' as http;
-import 'package:off_top_mobile/components/footer/bottomNavigationTabs.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:google_sign_in/google_sign_in.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
-
-class LoginPage extends StatefulWidget {
-  const LoginPage({Key key}) : super(key: key);
-
-  @override
-  _LoginPageState createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  int userId;
-  String userEmail;
-  String name;
-  bool showLoading = false;
-
-  final GoogleSignIn _googleSignIn = GoogleSignIn();
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  Future<FirebaseUser> _handleSignIn() async {
-    FirebaseUser user;
-    final bool isSignedIn = await _googleSignIn.isSignedIn();
-    if (isSignedIn) {
-      user = await _auth.currentUser();
-    } else {
-      final GoogleSignInAccount googleUser = await _googleSignIn.signIn();
-      final GoogleSignInAuthentication googleAuth =
-          await googleUser.authentication;
-
-      final AuthCredential credential = GoogleAuthProvider.getCredential(
-          idToken: googleAuth.idToken, accessToken: googleAuth.accessToken);
-
-      user = (await _auth.signInWithCredential(credential)).user;
-    }
-    return user;
-  }
-
-  Future<void> onGoogleSignIn(BuildContext context) async {
-    final FirebaseUser user = await _handleSignIn();
-    userEmail = user.email;
-    name = user.displayName;
-  }
-
-  Future<void> makeLoginRequest() async {
-    final String userEmail = this.userEmail;
-    final String url = 'http://localhost:9000/user/$userEmail';
-    //String url = 'http://10.0.2.2:9000/user/${userEmail}/';
-    final http.Response response = await http.get(Uri.encodeFull(url),
-        headers: <String, String>{'Accept': 'application/json'});
-    final dynamic userData = json.decode(response.body);
-    final SharedPreferences prefs = await SharedPreferences.getInstance();
-    prefs.setString(
-      'name',
-      userData['firstName'].toString(),
-    );
-    setState(
-      () {
-        userId = int.parse(userData['Id'].toString());
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).backgroundColor,
-      body: ModalProgressHUD(
-        inAsyncCall: showLoading,
-        child: Container(
-          padding: const EdgeInsets.all(50),
-          child: Align(
-            alignment: Alignment.center,
-            child: FlatButton(
-              shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20)),
-              onPressed: () async {
-                try {
-                  setState(
-                    () {
-                      showLoading = true;
-                    },
-                  );
-                  await onGoogleSignIn(context);
-                  await makeLoginRequest();
-                  if (userId == null) {
-                    setState(() {
-                      const Center(
-                        child: CircularProgressIndicator(),
-                      );
-                    });
-                  }
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute<void>(
-                      builder: (BuildContext context) => BottomNavigationTabs(
-                        RecordingPage(userId: userId),
-                      ),
-                    ),
-                  );
-                  setState(() {
-                    showLoading = false;
-                  });
-                } catch (e) {
-                  setState(() {
-                    showLoading = false;
-                  });
-                }
-              },
-              color: Colors.lightBlue,
-              child: Padding(
-                padding: const EdgeInsets.all(10),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: <Widget>[
-                    Icon(Icons.account_circle, color: Colors.white),
-                    const SizedBox(width: 10),
-                    const Text('Sign in with Google!'),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-  }
-}
